requirements.txt
streamlit
openai
langchain
PyMuPDF
genanki
python-dotenv